% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setRproj.R
\name{setRproj}
\alias{setRproj}
\title{Title
funtion to set up a R project}
\usage{
setRproj(fd_name)
}
\arguments{
\item{fd_name}{working folder/project name of the project}
}
\value{
.Rproj and folder structure of the project
}
\description{
Title
funtion to set up a R project
}
\examples{
\dontrun{
setProj("/Users/newfolder")
}
}
